+++ cypress/support/e2e.js
@@ -1,0 +1,8 @@
+// Custom commands and global before/after hooks for Cypress
+defineCypressKeyboardTab();
+
+function defineCypressKeyboardTab() {
+  Cypress.Commands.add('tab', { prevSubject: 'element' }, (subject) => {
+    cy.wrap(subject).trigger('keydown', { keyCode: 9, which: 9 });
+  });
+}